requirements.txt
opencv-python
PyTurboJPEG
pigpio
pygame
flask